+++ src/test/java/org/morphix/convert/extras/ExcludedFieldsTest.java
@@ -130,6 +130,7 @@ class ExcludedFieldsTest {
 		assertThat(e.equals(null), equalTo(false));
 	}
 
+	@SuppressWarnings("unlikely-arg-type")
 	@Test
 	void shouldReturnFalseWhenComparingDifferentClassOnEquals() {
 		ExcludedFields e = ExcludedFields.exclude("b");