+++ src/test/java/org/morphix/reflection/ConstantPoolTest.java
@@ -107,8 +107,8 @@ class ConstantPoolTest {
 	}
 
 	@Test
-	@SuppressWarnings("unchecked")
 	void shouldDecrementIndexOnNextElement() {
+		@SuppressWarnings("unchecked")
 		ConstantPoolIterator<Member> iterator = (ConstantPoolIterator<Member>) constantPool.iterator();
 		int index = iterator.nextIndex();
 		int nextIndex = iterator.nextIndex();
@@ -126,9 +126,8 @@ class ConstantPoolTest {
 
 	@ParameterizedTest
 	@MethodSource("provideIndexesTestParameters")
-	@SuppressWarnings("unchecked")
 	void shouldVerifyAllIteratorIndexes(final int size, final int index, final boolean hasNext) {
-		ConstantPoolIterator<Member> iterator = (ConstantPoolIterator<Member>) constantPool.iterator();
+		Iterator<Member> iterator = constantPool.iterator();
 		Fields.IgnoreAccess.set(iterator, "size", size);
 		Fields.IgnoreAccess.set(iterator, "index", index);